+++ packages/zwave-js/src/lib/commandclass/DoorLockCC.ts
@@ -179,9 +179,10 @@ export class DoorLockCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			DoorLockCCCapabilitiesReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<DoorLockCCCapabilitiesReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, [
 			"autoRelockSupported",
 			"blockToBlockSupported",
@@ -208,9 +209,10 @@ export class DoorLockCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			DoorLockCCOperationReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<DoorLockCCOperationReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, [
 			"currentMode",
 			"targetMode",
@@ -271,9 +273,10 @@ export class DoorLockCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			DoorLockCCConfigurationReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<DoorLockCCConfigurationReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, [
 			"operationType",
 			"outsideHandlesCanOpenDoorConfiguration",