+++ packages/zwave-js/src/lib/node/Node.ts
@@ -2435,9 +2435,7 @@ version:               ${this.version}`;
 		this._firmwareUpdateStatus.abort = true;
 
 		try {
-			await this.driver.waitForCommand<
-				FirmwareUpdateMetaDataCCStatusReport
-			>(
+			await this.driver.waitForCommand<FirmwareUpdateMetaDataCCStatusReport>(
 				(cc) =>
 					cc.nodeId === this.nodeId &&
 					cc instanceof FirmwareUpdateMetaDataCCStatusReport &&
@@ -2664,9 +2662,7 @@ version:               ${this.version}`;
 
 	private async finishFirmwareUpdate(): Promise<void> {
 		try {
-			const report = await this.driver.waitForCommand<
-				FirmwareUpdateMetaDataCCStatusReport
-			>(
+			const report = await this.driver.waitForCommand<FirmwareUpdateMetaDataCCStatusReport>(
 				(cc) =>
 					cc.nodeId === this.nodeId &&
 					cc instanceof FirmwareUpdateMetaDataCCStatusReport,